requirements.txt
aiohttp=3.9.3=py38h336bac9_1
aiosignal=1.3.1=pyhd8ed1ab_0
annotated-types=0.6.0=pyhd8ed1ab_0
anyio=4.3.0=pyhd8ed1ab_0
async-timeout=4.0.3=pyhd8ed1ab_0
attrs=23.2.0=pyh71513ae_0
av=11.0.0=pypi_0
boto3=1.34.73=pyhd8ed1ab_0
botocore=1.34.73=pyge38_1234567_0
brotli-python=1.1.0=py38he333c0f_1
bzip2=1.0.8=h93a5062_5
ca-certificates=2024.2.2=hf0a4a13_0
certifi=2024.2.2=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
coloredlogs=15.0.1=pypi_0
ctranslate2=4.1.0=pypi_0
dataclasses-json=0.6.4=pyhd8ed1ab_0
decorator=4.4.2=pypi_0
distro=1.9.0=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
faster-whisper=1.0.1=pypi_0
filelock=3.13.4=pypi_0
flatbuffers=24.3.25=pypi_0
frozenlist=1.4.1=py38h336bac9_0
fsspec=2024.3.1=pypi_0
greenlet=3.0.3=py38hb9fa5a8_0
h11=0.14.0=pyhd8ed1ab_0
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
httpcore=1.0.5=pyhd8ed1ab_0
httpx=0.27.0=pyhd8ed1ab_0
huggingface-hub=0.22.2=pypi_0
humanfriendly=10.0=pypi_0
hyperframe=6.0.1=pyhd8ed1ab_0
idna=3.6=pyhd8ed1ab_0
imageio=2.34.0=pypi_0
imageio-ffmpeg=0.4.9=pypi_0
jmespath=1.0.1=pyhd8ed1ab_0
jsonpatch=1.33=pyhd8ed1ab_0
jsonpointer=2.4=py38h10201cd_3
langchain=0.1.16=pyhd8ed1ab_0
langchain-community=0.0.32=pyhd8ed1ab_0
langchain-core=0.1.42=pyhd8ed1ab_0
langchain-experimental=0.0.57=pypi_0
langchain-openai=0.1.3=pypi_0
langchain-text-splitters=0.0.1=pyhd8ed1ab_0
langsmith=0.1.45=pyhd8ed1ab_0
libblas=3.9.0=22_osxarm64_openblas
libcblas=3.9.0=22_osxarm64_openblas
libcxx=16.0.6=h4653b0c_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
liblapack=3.9.0=22_osxarm64_openblas
libopenblas=0.3.27=openmp_h6c19121_0
libsqlite=3.45.2=h091b4b1_0
libzlib=1.2.13=h53f4e23_5
llvm-openmp=18.1.3=hcd81f8e_0
marshmallow=3.21.1=pyhd8ed1ab_0
moviepy=1.0.3=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=py38hb991d35_0
mypy_extensions=1.0.0=pyha770c72_0
ncurses=6.4.20240210=h078ce10_0
numpy=1.24.4=py38ha84db1f_0
onnxruntime=1.17.1=pypi_0
openai=1.14.3=pyhd8ed1ab_0
openssl=3.2.1=h0d3ecfb_1
orjson=3.9.15=py38h51b07bf_0
packaging=23.2=pyhd8ed1ab_0
pillow=10.3.0=pypi_0
pip=24.0=pyhd8ed1ab_0
proglog=0.1.10=pypi_0
protobuf=5.26.1=pypi_0
pydantic=2.0.3=pyhd8ed1ab_1
pydantic-core=2.3.0=py38hd0c8013_0
pysocks=1.7.1=pyha2e5f31_6
python=3.8.19=h2469fbe_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-dotenv=1.0.1=pyhd8ed1ab_0
python_abi=3.8=4_cp38
pyyaml=6.0.1=py38hb192615_1
readline=8.2=h92ec313_1
regex=2023.12.25=pypi_0
requests=2.31.0=pyhd8ed1ab_0
s3transfer=0.10.1=pyhd8ed1ab_0
setuptools=69.2.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.1=pyhd8ed1ab_0
sqlalchemy=2.0.29=py38h336bac9_0
sympy=1.12=pypi_0
tenacity=8.2.3=pyhd8ed1ab_0
tiktoken=0.6.0=pypi_0
tk=8.6.13=h5083fa2_1
tokenizers=0.15.2=pypi_0
tqdm=4.66.2=pyhd8ed1ab_0
typing-extensions=4.10.0=hd8ed1ab_0
typing_extensions=4.10.0=pyha770c72_0
typing_inspect=0.9.0=pyhd8ed1ab_0
urllib3=1.26.18=pyhd8ed1ab_0
wheel=0.43.0=pyhd8ed1ab_1
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
yarl=1.9.4=py38h336bac9_0
